INFRA-1083: change prefetch count to int

See merge request globalid/common/java/globalid-common!11

+++ src/main/java/id/global/common/annotations/amqp/MessageHandler.java
@@ -65,5 +65,5 @@ public @interface MessageHandler {
      * 
      * @return number of messages to fetch
      */
-    long prefetchCount() default 1;
+    int prefetchCount() default 1;
 }
